Update combat-trainer.lic
The shield rush mechanic has a bug (on DR's side), and as a result combat-trainer can get stuck in a short loop. It can only happens if you're hunting in the same room as someone else. I'm not sure if combat pets can trigger it. It happens pretty frequently "naturally" when using RUSH in a room with another person. 

I can't reproduce it reliably, but it can happen after either person kills something, and you attempt to RUSH to the next target. In some circumstances, it will tell you you're already at melee even though you're definitely not. And no amount of retreating will put you at rush-able distance. combat-trainer will get stuck in a loop of you retreating, removing your shield, trying and failing to rush, until the other person kills that target.

I filed a DR bug report, but I don't know if it'll be fixed since I couldn't explain how to reproduce it consistently.

So, this edit will retry RUSHing once when (allegedly) not at melee, but then set the cooldown timer so that other combat actions are able to reset the engagement. It can be reverted if/when the DR bug is fixed.